Public API

Contents

Public API

Contents

Index

Public interface

Index

GershgorinDiscs.Disc

GershgorinDiscs.GershgorinDisc

GershgorinDiscs.GershgorinDisc

Base.in

Base.in

Base.isapprox

GershgorinDiscs.eigvals_extrema

GershgorinDiscs.is_center_real

GershgorinDiscs.is_concentric

GershgorinDiscs.list_discs

Public interface

To utilize the public interface, first import the required packages:

julia> using GershgorinDiscs

Types

GershgorinDiscs.Disc — Type

A alias to GershgorinDisc.

GershgorinDiscs.GershgorinDisc — Type

GershgorinDisc{T}

Represent a Gershgorin disc in the complex plane associated with a matrix.

Arguments

center: a tuple (real_part, imaginary_part) representing the center of the disc.

radius: a non-negative real number representing the radius of the disc.

GershgorinDiscs.GershgorinDisc — Method

GershgorinDisc(x::Number, radius)

Construct a GershgorinDisc from a real or complex number and a radius.

Functions

Base.in — Method

in(number, disc::GershgorinDisc)

Check if a number is within the Gershgorin disc.

Base.in — Method

in(a::GershgorinDisc, b::GershgorinDisc)

Check if one Gershgorin disc is within another Gershgorin disc.

Base.isapprox — Method

isapprox(a::GershgorinDisc, b::GershgorinDisc; kwargs...)

Check if two GershgorinDisc objects are approximately equal.

GershgorinDiscs.eigvals_extrema — Method

eigvals_extrema(A::AbstractMatrix)

Estimate the minimum and maximum eigenvalues of a square matrix A using Gershgorin circle theorem.

This function computes the Gershgorin discs for A and returns the smallest and largest values that any eigenvalue could have based on the discs.

GershgorinDiscs.is_center_real — Method

is_center_real(d::GershgorinDisc)

Check whether the center of the given GershgorinDisc is real.

GershgorinDiscs.is_concentric — Method

is_concentric(a::GershgorinDisc, b::GershgorinDisc)

Check whether two GershgorinDisc objects are concentric.

GershgorinDiscs.list_discs — Method

list_discs(A::AbstractMatrix)

Compute the Gershgorin discs for a square matrix.

Each disc has its center at a diagonal element, and its radius is the smaller value between the row and column sums for that element.

Arguments

A::AbstractMatrix: a square matrix (either real or complex).

Examples

julia> A = [4.0 1.0; 0.5 3.0]
2×2 Matrix{Float64}:
 4.0  1.0
 0.5  3.0

julia> list_discs(A)
2-element Vector{GershgorinDisc{Float64}}:
 GershgorinDisc{Float64}((4.0, 0.0), 0.5)
 GershgorinDisc{Float64}((3.0, 0.0), 0.5)